# openapi_gateway_tmp.yaml
swagger: "2.0"
info:
  title: "LearnSphere API Gateway"
  description: "Unified entry point for LearnSphere services"
  version: "1.0.0"
schemes:
  - "https"
paths:
  /api/courses:
    post:
      operationId: "createCourse"
      x-google-backend:
        address: "https://course-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"
  /api/courses/{course_id}:
    get:
      operationId: "getCourse"
      parameters:
        - name: "course_id"
          in: "path"
          required: true
          type: "string"
      x-google-backend:
        address: "https://course-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"
  /api/courses/{course_id}/modules:
    parameters:
      - name: "course_id"
        in: "path"
        required: true
        type: "string"
    post:
      operationId: "addModule"
      x-google-backend:
        address: "https://course-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"
    get:
      operationId: "listModules"
      x-google-backend:
        address: "https://course-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"
  /api/courses/{course_id}/materials:
    parameters:
      - name: "course_id"
        in: "path"
        required: true
        type: "string"
    post:
      operationId: "addMaterial"
      x-google-backend:
        address: "https://course-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"
    get:
      operationId: "listMaterials"
      x-google-backend:
        address: "https://course-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"

  /api/assignments:
    post:
      operationId: "createAssignment"
      x-google-backend:
        address: "https://assessment-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"
  /api/assignments/{assignment_id}:
    get:
      operationId: "getAssignment"
      parameters:
        - name: "assignment_id"
          in: "path"
          required: true
          type: "string"
      x-google-backend:
        address: "https://assessment-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"

  /api/submissions:
    post:
      operationId: "submitAnswer"
      x-google-backend:
        address: "https://assessment-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"
  /api/submissions/{submission_id}:
    get:
      operationId: "getSubmission"
      parameters:
        - name: "submission_id"
          in: "path"
          required: true
          type: "string"
      x-google-backend:
        address: "https://assessment-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"
  /api/submissions/{submission_id}/grade:
    get:
      operationId: "getGrade"
      parameters:
        - name: "submission_id"
          in: "path"
          required: true
          type: "string"
      x-google-backend:
        address: "https://assessment-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"

  /api/profiles/{learner_id}:
    get:
      operationId: "getProfile"
      parameters:
        - name: "learner_id"
          in: "path"
          required: true
          type: "string"
      x-google-backend:
        address: "https://learning-profile-service-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"

  /api/agent/recommend:
    post:
      operationId: "recommend"
      x-google-backend:
        address: "https://recommender-agent-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"

  /api/agent/grade-submission:
    post:
      operationId: "gradeSubmission"
      x-google-backend:
        address: "https://auto-grading-agent-p2nkxa42qa-uc.a.run.app/"
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        200:
          description: "OK"